+++ lib/pusher-fake/configuration.rb
@@ -47,7 +47,8 @@ module PusherFake
     def to_options(options = {})
       options.merge(
         wsHost: socket_options[:host],
-        wsPort: socket_options[:port]
+        wsPort: socket_options[:port],
+        enabledTransports: ["ws"]
       )
     end